requirements.txt
alpaca-py>=0.18.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
eventlet>=0.35.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
ta>=0.11.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
pytz>=2024.1
aiohttp>=3.9.0
websocket-client>=1.7.0
scipy>=1.11.0
